Add collect, from Patrick Walton, to run an iterator and collect the results in a table.
As a result, rewrite all the table functionals to be iterator
functionals (in base) and reimplement the list functionals in terms of
them. Add two iterators for lists, elems and relems, that return only
the elements and not the indices, in order to implement the list
functionals.

A couple of old fixes where the Lua 5.1 table count is used (#).

Simplify ripairs slightly.

Fix a comment typo.