string_ext: fix a bad assumption in a spec example.
Running string.format in the expectation `should` uses the core
Lua string.format which doesn't prettify tables, but we're
comparing it to the string_ext `..` operator which uses the std
enhanced string.format, and that *does* prettify tables.
* specs/string_ext_spec.yaml (stringifies non-string arguments):
Manually expand the the stringified table in should to match the
std prettified table stringification.