configury: install correctly with configure or luarocks.
For the classic './configure; make; make install' to work we need
to install to '$luadir' as discovered by ax_lua.m4.  But we also
need to install directly to luarocks '$(LUADIR)' otherwise the
stdlib source files are not copied onto the LUA_PATH properly by
'luarocks make stdlib-*-1.rockspec'.
* mkrockspecs.lua: Move from here...
* build-aux/mkrockspecs: ...to here.
* GNUmakefile (rockspecs): Adjust accordingly.
* build-aux/mkrockspecs: Synchronise with the version that loads
a template, interpolates the variables and writes out the result.
* rockspecs.lua (build.install_command): Pass luadir to ensure
installation to the correct directory for luarocks.
* stdlib.rockspec.in: Remove.
* bootstrap: New script to call autoreconf for compatibility with
new build-aux/mkrockspecs.
* GNUmakefile (Makefile.in): Use it.