std: don't overwrite lua library tables.
Overwriting a global library table causes hard-to-debug problems
when the C part of Lua is holding references to the original
(overwritten) table.
* std/std.lua.in: When injecting new entry points into a core
library, do it an entry at a time to avoid breaking references to
the original address in the core.  This fixes a problem with the
`package.path` and `package.cpath` being lost when set from
LUA_INIT, among many other potential issues.