string: display prettyprinted table keys in table.sort order.
When comparing actual and expected output in unit tests, or with
Specl matchers, it's quite difficult to write robust examples when
the output of prettyprint is randomised for tables.  This commit
ensures that table keys are always output in the same order.
* specs/string_spec.yaml (prettytostring): Examples of desired
behaviour for string.prettytostring.
* lib/std/string.lua (render): Use a sorted list of keys to
determine the order in which they are output.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>