list: std.list returns the List prototype object.
Removing the separation between List prototype and the module
table that used to wrap it, for consistency with the other
Object specialisations.  We break backwards compatibilitf here
though because the argument ordering used to be different for
several functions depending on whether they were called through
the module table, or as List methods.  The List method ordering
(i.e. list argument first) wins out here, because that is a
prerequisite of being able to call object methods with the Lua
: syntax.
* specs/list_spec.yaml: Remove examples of constructing a List
from the module table, and with the `new` command.
* lib/std/list.lua: Much simplified by removal of the module
table.
(filter, foldl, foldr, index_key, index_value, map, map_with)
(project, shape, zip_with): The list operand is always first
now, whether called with a `.' or a `:'.
Adjust all callers.
(new): Remove.  Adjust all callers.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>