doc: separate camelCaseCompat methods to avoid LDoc warnings.
Using the @export feature of LDoc provides the flexibility to
change the exported functions from a single table, but we also
have to be careful not to sweep up undocumented access points
in the same table, otherwise LDoc correctly warns us that those
functions have no doc-comments.
* ext/std/getopt.lua (getOpt, processArgs, usageInfo): Move
to a separate table, and merge back into the module table before
return.
* ext/std/io.lua (processFiles): Likewise.
* ext/std/list.lua (indexKey, indexValue, mapWith, zipWith):
Likewise.
* ext/std/string.lua (escapePattern, escapeShell, ordinalSuffix):
Likewise.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>