alien: remove surplus std.alien module.
* lib/std/array.lua: Factor away dependencies on std.alien.
(__call): Instead of an _init call built for container.__call,
we now use a custom __call metamethod that returns a base_array
object that is optimized for Lua table buffers when the element
type is not recognized (either because there is no alien module
installed, or it does not support buffers of the given type), or
else builds an alien.buffer optimised object and returns that.
* lib/std/alien.lua, specs/alien_spec.yaml: Remove.
* build-aux/config.ld.in (file): Remove lib/std/alien.lua.
* local.mk (dist_luastd_DATA): Likewise.
(dist_modules_DATA): Remove doc/modules/std.alien.html.
* specs/specs.mk (specl_SPECS): Remove specs/alien_spec.yaml.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>