refactor: move argscheck et.al. into std.base module.
Unfortunately, debug.say is a very high level function that pulls
in a lot of other modules, modules that we'd like to be able to
add debug.argscheck to... since argscheck has almost no pre-
requisites, move it into the base module where it can then be
available everywhere else.
* lib/std/container.lua (prototype): Move from here...
* lib/std/base.lua (prototype): ...to here.
* lib/std/debug.lua (concat, argerror, argcheck, argscheck):
Move from here...
* lib/std/base.lua (concat, argerror, argcheck, argscheck): ...to
here.
Be sure to disable argument checking if _DEBUG or _DEBUG.argcheck
are false.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>