debug: revert export of argerror, argcheck and argscheck.
Annoyingly, Lua 5.1 misses an obvious tail call elimination when
_DEBUG.argcheck is set, so the deep call to error gets the wrong
level, and reports argument errors in the wrong functions!! Rather
than uglify the code to remove the tail-calls and do a recount,
or add a fudge factor when Lua 5.1 is detected, it's cleaner to
remove the argchecking wrappers of the 3 affected functions -- at
least until we're ready to drop Lua 5.1 support entirely.
* specs/debug_spec.yaml (argerror, argcheck, argscheck): Mark the
argument checking behaviours as pending.
* lib/std/debug.lua (argerror, argcheck, argscheck): Comment out
the argument checking wrappers, and call the bare functions.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>