debug: finish support for nil arguments in exported functions.
Because luajit (legitimately) stops counting on the first nil
value in a list, where Lua 5.1 and 5.2 keep counting, we have
to do our own size calculations on argument vectors to be
consistent.
* lib/std/base.lua (olen): Return the largest integer key from a
table.
(match, export): Use it to ignore `nil` elements in the argument
list when counting the number of arguments.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>