set: check for strict Set type arguments.
Since we have type checking to help maintain correctness, don't
undermine ourselves and add complexity by adding type coercions
and looser type safety.  When we're using Set's, passing a table
instead of a Set is probably an error, so treat it as one!
* specs/set_spec.yaml: Remove specifications for type coercions.
* lib/std/set.lua: Remove type coercions and require strict Set
arguments everywhere.
Add type checks to metamethods.
* NEWS: Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>